try golang module
