+++ api/main.go
@@ -1,51 +1,42 @@
 package main
 
 import (
-	"bytes"
 	"database/sql"
-	"fmt"
-	"log"
-	"net/http"
 	"os"
 	"time"
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/google/uuid"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awsutil"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3"
 
 	"github.com/joho/godotenv"
+
+	"github.com/jpskgc/article/api/controller"
 )
 
-type Article struct {
-	ID         int         `json:"id"`
-	UUID       string      `json:"uuid"`
-	TITLE      string      `json:"title"`
-	CONTENT    string      `json:"content"`
-	IMAGENAMES []ImageName `json:"imageNames"`
-}
+// type Article struct {
+// 	ID         int         `json:"id"`
+// 	UUID       string      `json:"uuid"`
+// 	TITLE      string      `json:"title"`
+// 	CONTENT    string      `json:"content"`
+// 	IMAGENAMES []ImageName `json:"imageNames"`
+// }
 
-type ImageName struct {
-	NAME string `json:"name"`
-}
+// type ImageName struct {
+// 	NAME string `json:"name"`
+// }
 
-type ImageData struct {
-	ARTICLEUUID string      `json:"articleUUID"`
-	IMAGENAMES  []ImageName `json:"imageNames"`
-}
+// type ImageData struct {
+// 	ARTICLEUUID string      `json:"articleUUID"`
+// 	IMAGENAMES  []ImageName `json:"imageNames"`
+// }
 
-type Param struct {
-	Bucket      string
-	Key         string
-	Expires     string
-	ContentType string
-}
+// type Param struct {
+// 	Bucket      string
+// 	Key         string
+// 	Expires     string
+// 	ContentType string
+// }
 
 func main() {
 	err := godotenv.Load()
@@ -100,134 +91,139 @@ func main() {
 	api := router.Group("/api")
 	{
 		api.GET("/articles", func(c *gin.Context) {
-			var articles []Article
-			results, err := db.Query("SELECT * FROM articles")
-			if err != nil {
-				panic(err.Error())
-			}
-			article := Article{}
-			for results.Next() {
-				err = results.Scan(&article.ID, &article.UUID, &article.TITLE, &article.CONTENT)
-				if err != nil {
-					panic(err.Error())
-				} else {
-					articles = append(articles, article)
-				}
-			}
-			c.JSON(http.StatusOK, articles)
+			controller.GetArticleController(c, db)
+			// var articles []Article
+			// results, err := db.Query("SELECT * FROM articles")
+			// if err != nil {
+			// 	panic(err.Error())
+			// }
+			// article := Article{}
+			// for results.Next() {
+			// 	err = results.Scan(&article.ID, &article.UUID, &article.TITLE, &article.CONTENT)
+			// 	if err != nil {
+			// 		panic(err.Error())
+			// 	} else {
+			// 		articles = append(articles, article)
+			// 	}
+			// }
+			// c.JSON(http.StatusOK, articles)
 		})
 		api.GET("/article/:id", func(c *gin.Context) {
-			id := c.Params.ByName("id")
-			article := Article{}
-			errArticle := db.QueryRow("SELECT * FROM articles WHERE id = ?", id).Scan(&article.ID, &article.UUID, &article.TITLE, &article.CONTENT)
-			if errArticle != nil {
-				panic(errArticle.Error())
-			}
-			rows, errImage := db.Query("SELECT image_name FROM images WHERE article_uuid  = ?", article.UUID)
-			if errImage != nil {
-				panic(errImage.Error())
-			}
-			for rows.Next() {
-				imageName := ImageName{}
-				err = rows.Scan(&imageName.NAME)
-				article.IMAGENAMES = append(article.IMAGENAMES, imageName)
-			}
-			c.JSON(http.StatusOK, article)
+			controller.GetSingleArticleController(c, db)
+			// id := c.Params.ByName("id")
+			// article := Article{}
+			// errArticle := db.QueryRow("SELECT * FROM articles WHERE id = ?", id).Scan(&article.ID, &article.UUID, &article.TITLE, &article.CONTENT)
+			// if errArticle != nil {
+			// 	panic(errArticle.Error())
+			// }
+			// rows, errImage := db.Query("SELECT image_name FROM images WHERE article_uuid  = ?", article.UUID)
+			// if errImage != nil {
+			// 	panic(errImage.Error())
+			// }
+			// for rows.Next() {
+			// 	imageName := ImageName{}
+			// 	err = rows.Scan(&imageName.NAME)
+			// 	article.IMAGENAMES = append(article.IMAGENAMES, imageName)
+			// }
+			// c.JSON(http.StatusOK, article)
 		})
 		api.GET("/delete/:id", func(c *gin.Context) {
-			id := c.Params.ByName("id")
-			_, err = db.Exec("DELETE FROM articles WHERE id= ?", id)
-			if err != nil {
-				log.Fatalf("db.Exec(): %s\n", err)
-			}
-			c.JSON(http.StatusOK, gin.H{"status": "ok"})
+			controller.DeleteArticleController(c, db)
+			// id := c.Params.ByName("id")
+			// _, err = db.Exec("DELETE FROM articles WHERE id= ?", id)
+			// if err != nil {
+			// 	log.Fatalf("db.Exec(): %s\n", err)
+			// }
+			// c.JSON(http.StatusOK, gin.H{"status": "ok"})
 		})
 		api.POST("/post", func(c *gin.Context) {
-			u, err := uuid.NewRandom()
-			if err != nil {
-				fmt.Println(err)
-				return
-			}
-			uu := u.String()
-			var article Article
-			c.BindJSON(&article)
-			ins, err := db.Prepare("INSERT INTO articles(uuid, title,content) VALUES(?,?,?)")
-			if err != nil {
-				log.Fatal(err)
-			}
-			ins.Exec(uu, article.TITLE, article.CONTENT)
+			controller.PostController(c, db)
+			// u, err := uuid.NewRandom()
+			// if err != nil {
+			// 	fmt.Println(err)
+			// 	return
+			// }
+			// uu := u.String()
+			// var article Article
+			// c.BindJSON(&article)
+			// ins, err := db.Prepare("INSERT INTO articles(uuid, title,content) VALUES(?,?,?)")
+			// if err != nil {
+			// 	log.Fatal(err)
+			// }
+			// ins.Exec(uu, article.TITLE, article.CONTENT)
 
-			c.JSON(http.StatusOK, gin.H{"uuid": uu})
+			// c.JSON(http.StatusOK, gin.H{"uuid": uu})
 
 		})
 		api.POST("/post/image", func(c *gin.Context) {
+			controller.PostImageController(c)
+			// creds := credentials.NewStaticCredentials(os.Getenv("AWS_ACCESS_KEY_ID"), os.Getenv("AWS_SECRET_ACCESS_KEY"), "")
 
-			creds := credentials.NewStaticCredentials(os.Getenv("AWS_ACCESS_KEY_ID"), os.Getenv("AWS_SECRET_ACCESS_KEY"), "")
+			// cfg := aws.NewConfig().WithRegion("ap-northeast-1").WithCredentials(creds)
+			// svc := s3.New(session.New(), cfg)
 
-			cfg := aws.NewConfig().WithRegion("ap-northeast-1").WithCredentials(creds)
-			svc := s3.New(session.New(), cfg)
+			// form, _ := c.MultipartForm()
 
-			form, _ := c.MultipartForm()
+			// files := form.File["images[]"]
 
-			files := form.File["images[]"]
+			// var imageNames []ImageName
+			// imageName := ImageName{}
 
-			var imageNames []ImageName
-			imageName := ImageName{}
+			// for _, file := range files {
 
-			for _, file := range files {
+			// 	f, err := file.Open()
 
-				f, err := file.Open()
+			// 	if err != nil {
+			// 		log.Println(err)
+			// 	}
 
-				if err != nil {
-					log.Println(err)
-				}
+			// 	defer f.Close()
 
-				defer f.Close()
+			// 	size := file.Size
+			// 	buffer := make([]byte, size)
 
-				size := file.Size
-				buffer := make([]byte, size)
+			// 	u, err := uuid.NewRandom()
+			// 	if err != nil {
+			// 		fmt.Println(err)
+			// 		return
+			// 	}
+			// 	uu := u.String()
 
-				u, err := uuid.NewRandom()
-				if err != nil {
-					fmt.Println(err)
-					return
-				}
-				uu := u.String()
+			// 	f.Read(buffer)
+			// 	fileBytes := bytes.NewReader(buffer)
+			// 	fileType := http.DetectContentType(buffer)
+			// 	path := "/media/" + uu
+			// 	params := &s3.PutObjectInput{
+			// 		Bucket:        aws.String("article-s3-jpskgc"),
+			// 		Key:           aws.String(path),
+			// 		Body:          fileBytes,
+			// 		ContentLength: aws.Int64(size),
+			// 		ContentType:   aws.String(fileType),
+			// 	}
+			// 	resp, err := svc.PutObject(params)
 
-				f.Read(buffer)
-				fileBytes := bytes.NewReader(buffer)
-				fileType := http.DetectContentType(buffer)
-				path := "/media/" + uu
-				params := &s3.PutObjectInput{
-					Bucket:        aws.String("article-s3-jpskgc"),
-					Key:           aws.String(path),
-					Body:          fileBytes,
-					ContentLength: aws.Int64(size),
-					ContentType:   aws.String(fileType),
-				}
-				resp, err := svc.PutObject(params)
+			// 	fmt.Printf("response %s", awsutil.StringValue(resp))
 
-				fmt.Printf("response %s", awsutil.StringValue(resp))
+			// 	imageName.NAME = uu
 
-				imageName.NAME = uu
+			// 	imageNames = append(imageNames, imageName)
+			// }
 
-				imageNames = append(imageNames, imageName)
-			}
-
-			c.JSON(http.StatusOK, imageNames)
+			// c.JSON(http.StatusOK, imageNames)
 		})
 		api.POST("/post/image/db", func(c *gin.Context) {
-			var imageData ImageData
-			c.BindJSON(&imageData)
+			controller.PostImageToDBController(c, db)
+			// var imageData ImageData
+			// c.BindJSON(&imageData)
 
-			for _, imageName := range imageData.IMAGENAMES {
-				ins, err := db.Prepare("INSERT INTO images(article_uuid, image_name) VALUES(?,?)")
-				if err != nil {
-					log.Fatal(err)
-				}
-				ins.Exec(imageData.ARTICLEUUID, imageName.NAME)
-			}
-			c.JSON(http.StatusOK, gin.H{"status": "ok"})
+			// for _, imageName := range imageData.IMAGENAMES {
+			// 	ins, err := db.Prepare("INSERT INTO images(article_uuid, image_name) VALUES(?,?)")
+			// 	if err != nil {
+			// 		log.Fatal(err)
+			// 	}
+			// 	ins.Exec(imageData.ARTICLEUUID, imageName.NAME)
+			// }
+			// c.JSON(http.StatusOK, gin.H{"status": "ok"})
 		})
 	}
